Abstract Snowpipe specific error mapping to class to allow the cloudwatch logging and metric class to be generalised
