+++ aws/error-notification/lambda/cloudwatch_client.py
@@ -1,6 +1,5 @@
 import logging
 from json import loads, dumps
-from datetime import datetime
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
@@ -17,31 +16,6 @@ class CloudWatchObservability:
             "logGroupName": f"{self.event.get('TopicArn').split(':')[-1]}",
         }
         return cloudwatch_parameters
-
-    def create_error_map_from_event(self):
-        timestamp = self.event.get("Timestamp")
-        epoch_timestamp_in_milliseconds = (
-            convert_timestamp_to_epoch_time_in_milliseconds(timestamp)
-        )
-        message = loads(self.event.get("Message"))
-        error_type = message.get("messageType")
-        topic = self.event.get("TopicArn")
-        pipe_name = message.get("pipeName")
-        table_name = message.get("tableName")
-        source = message.get("stageLocation")
-        message = message.get("messages")
-        snowpipe_error_map = {
-            "error_type": error_type,
-            "topic": topic,
-            "timestamp": epoch_timestamp_in_milliseconds,
-            "facts": {
-                "pipe_name": pipe_name,
-                "table_name": table_name,
-                "source": source,
-                "message": message,
-            },
-        }
-        return snowpipe_error_map
 
     def create_error_log_group(self, parameters):
         try:
@@ -71,13 +45,13 @@ class CloudWatchObservability:
         if stream_description.get("logStreams"):
             return stream_description.get("logStreams")[0].get("uploadSequenceToken")
 
-    def put_error_log_event(self, parameters, snowpipe_error_map):
+    def put_error_log_event(self, parameters, resource_error_map):
         parameters.update(
             {
                 "logEvents": [
                     {
-                        "timestamp": snowpipe_error_map.get("timestamp"),
-                        "message": dumps(snowpipe_error_map),
+                        "timestamp": resource_error_map.get("timestamp"),
+                        "message": dumps(resource_error_map),
                     }
                 ]
             }
@@ -91,8 +65,8 @@ class CloudWatchObservability:
             if error_code != "DataAlreadyAcceptedException":
                 raise exception
 
-    def put_metric(self, parameters, snowpipe_error_map):
-        pipe_name_in_error = snowpipe_error_map.get("facts").get("pipe_name")
+    def put_metric(self, parameters, resource_error_map):
+        pipe_name_in_error = resource_error_map.get("facts").get("pipe_name")
         response = self.cloudwatch_client.put_metric_data(
             Namespace="snowpipe-alerts",
             MetricData=[
@@ -119,10 +93,3 @@ class CloudWatchObservability:
             ],
         )
         return response
-        # logger.info(f"Number of new objects pushed to CloudWatch\n")
-
-
-def convert_timestamp_to_epoch_time_in_milliseconds(timestamp):
-    epoch_timestamp = datetime.strptime(timestamp, "%Y-%m-%dT%H:%M:%S.%fZ").timestamp()
-    epoch_timestamp_in_milliseconds = int(epoch_timestamp * 1000)
-    return epoch_timestamp_in_milliseconds
